Fix some links

+++ rebuild-old-html.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
-for x in *.org; do
+for x in $(find blog -name '*.org'); do
     if test ! -e $x; then
         continue
     fi
-    org=$(readlink $x);
+    org=$(readlink -f $x);
     html=${org/%.org/.html};
-    cp "$html" .;
-    perl -npe 's!(href\s*=\s*.|src\s*=\s*.)\.\.[./]*!$1!ig' -i $(basename $html);
+    if test -e $(basename $html); then
+        cp "$html" .;
+        perl -npe 's!(href\s*=\s*.|src\s*=\s*.)\.\.[./]*!$1!ig' -i $(basename $html);
+    fi
 done
